--- src/app/step/Step.ts
@@ -1,4 +1,0 @@
-export class Step {
-  title: string;
-  content: string;
-}
