+++ src/app/step/step-component/step.component.scss
