+++ src/app/step/terminal/terminal.component.html
