+++ src/app/step/terminal/terminal.component.scss
